deploy-hook/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.28.0
pyyaml==6.0.2
